Build a Library Codeacademy
Let’s start by making a parent class for our Book, CD, and Movie classes. Create an empty class called Media.
Inside the Media class, create an empty constructor() method that takes one parameter. This argument will set the one property that is in all three subclasses of Media, and does not have a default value.
Inside the constructor, set the values for Media properties that Book, CD, and Movie share. Prepend all of the property names with an underscore (_).
Create getters for the title, isCheckedOut, and ratings properties. Each getter should return the value saved to the matching property.
Create a setter for the isCheckedOut property.
Under your getters, create a method named toggleCheckOutStatus that changes the value saved to the _isCheckedOut property. If the value is true, then change it to false. If the value is false, then change it to true.
Under .toggleCheckOutStatus(), create a method named getAverageRating. Return the average value of the ratings array. Use the reduce method to find the sum of the ratings array. Divide this sum by the length of the ratings array, and return the result.
Let’s add a method named addRating that accepts one argument and uses .push() to add it to the end of the ratings array.
Next, we’’ll build a Book class that extends Media. If you feel comfortable building the Book class on your own, give it a shot. If not, use the steps below to help you along the way.
Whether you want to follow the steps or not, use the list of properties, getters, and methods as a reference.
Book
Properties: author (string), title (string), pages (number), isCheckedOut (boolean, initially false), and ratings (array, initially empty).
Getters: all properties have a getter
Methods: .getAverageRating(), .toggleCheckOutStatus(), and .addRating()
Create an empty Book class that extends Media.
Inside the Book class, create a constructor that accepts three arguments. These arguments are used to set properties that do not have default values.
Call super on the first line of the Books‘s constructor method. Pass it any arguments that the parent constructor uses.
Use the remaining arguments to set the author and pages properties in Book.
Since our Book class inherits Media‘s properties and getters, we only need to create two new getters in the Book class. Add two new getters to the Book class. Each getter should return the value saved to its matching property.
Let’s see if you can create an entire Movie class using only the property, getter, and method specifications below:
Movie
properties: director (string), title (string), runTime (number), isCheckedOut (boolean, initially false), and ratings (array, initially empty)
getters: all properties have a getter
methods: .getAverageRating(), .toggleCheckOutStatus(), and .addRating()
Create a Book instance with the following properties:
Author: 'Bill Bryson'
Title: 'A Short History of Nearly Everything'
pages: 544
Save the instance to a constant variable named historyOfEverything.
Call .toggleCheckOutStatus() on the historyOfEverything instance.
Log the value saved to the isCheckedOut property in the historyOfEverything instance.
Call .addRating() three times on historyOfEverything with inputs of 4, 5, and 5.
Call .getAverageRating() on historyOfEverything. Log the result to the console.
Create a Movie instance with the following properties:
Director: 'Jan de Bont'
Title: 'Speed'
Runtime: 116
Save the instance to a constant variable named speed.
Call .toggleCheckOutStatus() on the speed instance.
Log the value saved to the isCheckedOut property in the speed instance.
Call .addRating() three times on speed with inputs of 1, 1, and 5.
Call .getAverageRating() on speed. Log the result to the console.